+++ integration_tests/test_audit_critical_dxf_files.py
@@ -6,10 +6,13 @@ import ezdxf
 
 BASEDIR = 'integration_tests' if os.path.exists('integration_tests') else '.'
 DATADIR = 'data'
-COLDFIRE = r"D:\Source\dxftest\CADKitSamples\kit-dev-coldfire-xilinx_5213.dxf"
+COLDFIRE = os.path.join(
+    ezdxf.EZDXF_TEST_FILES,
+    "CADKitSamples/kit-dev-coldfire-xilinx_5213.dxf")
 
 
-@pytest.mark.skipif(not os.path.exists(COLDFIRE), reason='test data not present')
+@pytest.mark.skipif(not os.path.exists(COLDFIRE),
+                    reason='test data not present')
 def test_kit_dev_coldfire():
     doc = ezdxf.readfile(COLDFIRE)
     auditor = doc.audit()