+++ integration_tests/test_write_without_handles.py
@@ -14,7 +14,7 @@ DATADIR = 'data'
 def filename(request):
     filename = os.path.join(BASEDIR, DATADIR, request.param)
     if not os.path.exists(filename):
-        pytest.skip('File {} not found.'.format(filename))
+        pytest.skip(f'File {filename} not found.')
     return filename
 
 
@@ -44,5 +44,7 @@ def test_write_R12_without_handles(filename, tmpdir):
                 with pytest.raises(ezdxf.DXFValueError):  # has no handles
                     entity.get_handle()
             else:  # special DIMSTYLE entity
-                assert len(entity.find_all(105)) == 0, 'remove handle group code 105'
-                assert len(entity.find_all(5)) == 1, 'do not remove group code 5'
+                assert len(
+                    entity.find_all(105)) == 0, 'remove handle group code 105'
+                assert len(
+                    entity.find_all(5)) == 1, 'do not remove group code 5'