+++ NEWS.md
@@ -56,7 +56,8 @@ Version 0.14a4 - dev
 - NEW: `DXFEntity.new_extension_dict()`, create explicit a new extension dictionary
 - NEW: `ezdxf.reorder`, support module to implement modified entities redraw order
 - NEW: `ezdxf.recover`, module to load DXF Documents with structural flaws
-- NEW: get DXF test file path from environment variable `EZDXF_TEST_FILES`
+- NEW: get DXF test file path from environment variable `EZDXF_TEST_FILES`, 
+  imported automatically as `ezdxf.EZDXF_TEST_FILES`
 - CHANGE: `DXFEntity.get_extension_dict()`, raises `AttributeError` if entity
   has no extension dictionary 
 - CHANGE: `DXFEntity.has_extension_dict` is now a property not a method